+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.4] - 2020-08-26
+
 ### Fixed
 
 - Fix cluster status is not updated during cluster upgrade
@@ -127,7 +129,8 @@ be edited by users.
 
 
 
-[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v2.3.3...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v2.3.4...HEAD
+[2.3.4]: https://github.com/giantswarm/cluster-operator/compare/v2.3.3...v2.3.4
 [2.3.3]: https://github.com/giantswarm/cluster-operator/compare/v2.3.2...v2.3.3
 [2.3.2]: https://github.com/giantswarm/cluster-operator/compare/v2.3.1...v2.3.2
 [2.3.1]: https://github.com/giantswarm/cluster-operator/compare/v2.3.0...v2.3.1